discordbot/requirements.txt
python-dotenv
discord.py
redis
